`Removed outdated code: nnnc_app.py, Dockerfile, and GitHub workflows`
