--- .devcontainer/dockerfile.txt
@@ -1,46 +1,0 @@
-# Use lightweight Linux base
-FROM ubuntu:22.04
-
-# Set environment variables
-ENV DEBIAN_FRONTEND=noninteractive
-ENV MISE_HOME=/root/.local/share/mise
-ENV PATH=$MISE_HOME/bin:$PATH
-
-# Install dependencies
-RUN apt-get update && apt-get install -y \
-    curl \
-    git \
-    build-essential \
-    zstd \
-    wget \
-    tar \
-    python3-distutils \
-    ca-certificates \
-    && rm -rf /var/lib/apt/lists/*
-
-# Install Mise
-RUN curl -sSf https://github.com/albiemc1303/mise/releases/latest/download/mise-linux -o /usr/local/bin/mise \
-    && chmod +x /usr/local/bin/mise \
-    && mise self-update
-
-# Ensure any broken Python archives are cleared
-RUN rm -rf $MISE_HOME/downloads/python/3.12.12
-
-# Force Mise to use precompiled Python (with Zstandard installed)
-RUN mise use -g python@3.12
-
-# Set working directory
-WORKDIR /app
-
-# Copy and install Python dependencies
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Copy application code
-COPY . .
-
-# Expose port for FastAPI
-EXPOSE 8080
-
-# Command to run the Provenance Engine
-CMD ["uvicorn", "app:app", "--host", "0.0.0.0", "--port", "8080"]
