+++ .dockerignore
@@ -1,0 +1,5 @@
+fly.toml
+.git/
+__pycache__/
+.envrc
+.venv/
