+++ provenance_engine/dockerfile.txt
@@ -1,0 +1,22 @@
+
+# Use official Python image
+FROM python:3.11-slim
+
+# Set working directory
+WORKDIR /app
+
+# Copy requirements and install
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Copy application code
+COPY provenance_engine/app.py .
+
+# Environment variable (override with Fly secret)
+ENV PROVENANCE_API_TOKEN=changeme
+
+# Expose port
+EXPOSE 8080
+
+# Run FastAPI
+CMD ["uvicorn", "app:app", "--host", "0.0.0.0", "--port", "8080"]
